+++ com.unity.cinemachine/Editor/PropertyDrawers/FoldoutWithEnabledButtonPropertyDrawer.cs
@@ -75,7 +75,7 @@ namespace Unity.Cinemachine.Editor
 
             var row = new InspectorUtility.LabeledRow(preferredLabel, enabledProp.tooltip);
             var toggle = row.Contents.AddChild(new Toggle("") 
-                { style = { flexGrow = 0, marginTop = 2, marginLeft = 3, alignSelf = Align.Center }});
+                { style = { flexGrow = 0, marginTop = 3, marginBottom = 3, marginLeft = 3, alignSelf = Align.Center }});
             toggle.BindProperty(enabledProp);
 
             Label disabledText = null;
@@ -86,7 +86,7 @@ namespace Unity.Cinemachine.Editor
             var childLabel = row.Contents.AddChild(new Label(childProperty.displayName)
                 { style = { flexGrow = 0, marginLeft = 8, alignSelf = Align.Center }, tooltip = childProperty.tooltip});
             var childField = row.Contents.AddChild(new PropertyField(childProperty, "")
-                { style = { flexGrow = 1, marginTop = -1, marginLeft = 5, marginBottom = -1 }});
+                { style = { flexGrow = 1, marginTop = -1, marginLeft = 5, marginBottom = -1, marginRight = 1 }});
             childLabel.AddDelayedFriendlyPropertyDragger(childProperty, childField, (d) => d.CancelDelayedWhenDragging = true);
             childField.RemoveFromClassList(InspectorUtility.AlignFieldClassName);