+++ com.unity.cinemachine/Tests/Editor/BlendManagerTests.cs
@@ -29,6 +29,7 @@ namespace Unity.Cinemachine.Tests.Editor
         FakeMixer m_Mixer = new ("Mixer");
         FakeCamera m_Cam1 = new ("Cam1");
         FakeCamera m_Cam2 = new ("Cam2");
+        FakeCamera m_Cam3 = new ("Cam3");
 
         int m_ActivatedEventCount;
         int m_DeactivatedEventCount;
@@ -57,9 +58,18 @@ namespace Unity.Cinemachine.Tests.Editor
 
         void ResetCounters() => m_ActivatedEventCount = m_DeactivatedEventCount = m_BlendCreatedCount = m_BlendFinishedCount = 0;
 
-        void ProcessFrame(ICinemachineCamera cam, float deltaTime)
+        void Reset(float blendTime)
         {
-            m_BlendManager.UpdateRootFrame(m_Mixer, cam, Vector3.up, deltaTime);
+            m_BlendManager.LookupBlendDelegate = (outgoing, incoming)
+                => new (CinemachineBlendDefinition.Styles.EaseInOut, blendTime); // constant blend time
+            m_BlendManager.OnEnable();
+            ProcessFrame(null, 0.1f);
+            ResetCounters();
+        }
+
+        void ProcessFrame(ICinemachineCamera activeCam, float deltaTime)
+        {
+            m_BlendManager.UpdateRootFrame(m_Mixer, activeCam, Vector3.up, deltaTime);
             m_BlendManager.ComputeCurrentBlend();
             m_BlendManager.ProcessActiveCamera(m_Mixer, Vector3.up, deltaTime);
         }
@@ -67,11 +77,7 @@ namespace Unity.Cinemachine.Tests.Editor
         [Test]
         public void TestEvents()
         {
-            m_BlendManager.LookupBlendDelegate = (outgoing, incoming)
-                => new (CinemachineBlendDefinition.Styles.EaseInOut, 1); // constant blend time of 1
-
-            ResetCounters();
-            m_BlendManager.ResetRootFrame();
+            Reset(1); // constant blend time of 1
 
             // We should get an initial activation event, no blend
             ProcessFrame(m_Cam1, 0.1f);
@@ -106,6 +112,114 @@ namespace Unity.Cinemachine.Tests.Editor
             ProcessFrame(m_Cam2, 0.5f);
             Assert.AreEqual(2, m_ActivatedEventCount);
             Assert.AreEqual(1, m_DeactivatedEventCount);
+            Assert.AreEqual(1, m_BlendCreatedCount);
+            Assert.AreEqual(1, m_BlendFinishedCount);
+            Assert.That(m_BlendManager.IsBlending, Is.False);
+        }
+
+        [Test]
+        public void TestEventsNestedBlend()
+        {
+            Reset(1); // constant blend time of 1
+
+            // We should get an initial activation event, no blend
+            ProcessFrame(m_Cam1, 0.1f);
+            Assert.AreEqual(1, m_ActivatedEventCount);
+            Assert.AreEqual(0, m_DeactivatedEventCount);
+            Assert.AreEqual(0, m_BlendFinishedCount);
+            Assert.AreEqual(0, m_BlendCreatedCount);
+            Assert.That(m_BlendManager.IsBlending, Is.False);
+
+            ProcessFrame(m_Cam1, 0.1f);
+            Assert.AreEqual(1, m_ActivatedEventCount);
+            Assert.AreEqual(0, m_DeactivatedEventCount);
+            Assert.AreEqual(0, m_BlendCreatedCount);
+            Assert.AreEqual(0, m_BlendFinishedCount);
+            Assert.That(m_BlendManager.IsBlending, Is.False);
+
+            // Activate new camera, blend will take 1 sec
+            ProcessFrame(m_Cam2, 0.1f);
+            Assert.AreEqual(2, m_ActivatedEventCount);
+            Assert.AreEqual(0, m_DeactivatedEventCount);
+            Assert.AreEqual(1, m_BlendCreatedCount);
+            Assert.AreEqual(0, m_BlendFinishedCount);
+            Assert.That(m_BlendManager.IsBlending, Is.True);
+
+            ProcessFrame(m_Cam2, 0.5f);
+            Assert.AreEqual(2, m_ActivatedEventCount);
+            Assert.AreEqual(0, m_DeactivatedEventCount);
+            Assert.AreEqual(1, m_BlendCreatedCount);
+            Assert.AreEqual(0, m_BlendFinishedCount);
+            Assert.That(m_BlendManager.IsBlending, Is.True);
+
+            // Acivate new cam before old blend is finished
+            ProcessFrame(m_Cam3, 0.1f);
+            Assert.AreEqual(3, m_ActivatedEventCount);
+            Assert.AreEqual(0, m_DeactivatedEventCount);
+            Assert.AreEqual(2, m_BlendCreatedCount);
+            Assert.AreEqual(0, m_BlendFinishedCount);
+            Assert.That(m_BlendManager.IsBlending, Is.True);
+
+            // After first blend time has elapsed, check the counters
+            ProcessFrame(m_Cam3, 0.5f);
+            Assert.AreEqual(3, m_ActivatedEventCount);
+            Assert.AreEqual(1, m_DeactivatedEventCount);
+            Assert.AreEqual(2, m_BlendCreatedCount);
+            Assert.AreEqual(0, m_BlendFinishedCount); // blend was interrupted, never finished
+            Assert.That(m_BlendManager.IsBlending, Is.True);
+
+            // After second blend is finished, check the counters
+            ProcessFrame(m_Cam3, 0.5f);
+            Assert.AreEqual(3, m_ActivatedEventCount);
+            Assert.AreEqual(2, m_DeactivatedEventCount);
+            Assert.AreEqual(2, m_BlendCreatedCount);
+            Assert.AreEqual(1, m_BlendFinishedCount);
+            Assert.That(m_BlendManager.IsBlending, Is.False);
+        }
+
+        [Test]
+        public void TestEventsBlendToNestedBlend()
+        {
+            var customBlend = new NestedBlendSource(new CinemachineBlend()
+            {
+                CamA = m_Cam1,
+                CamB = m_Cam2,
+                BlendCurve = AnimationCurve.Linear(0, 0, 1, 1),
+                Duration = 1,
+                TimeInBlend = 0.1f
+            });
+
+            Reset(1); // constant blend time of 1
+
+            // We should get an initial activation event, no blend
+            ProcessFrame(m_Cam1, 0.1f);
+            Assert.AreEqual(1, m_ActivatedEventCount);
+            Assert.AreEqual(0, m_DeactivatedEventCount);
+            Assert.AreEqual(0, m_BlendFinishedCount);
+            Assert.AreEqual(0, m_BlendCreatedCount);
+            Assert.That(m_BlendManager.IsBlending, Is.False);
+
+            // Activate nested blend camera, blend will take 1 sec
+            ProcessFrame(customBlend, 0.1f);
+            Assert.AreEqual(2, m_ActivatedEventCount);
+            Assert.AreEqual(0, m_DeactivatedEventCount);
+            Assert.AreEqual(1, m_BlendCreatedCount);
+            Assert.AreEqual(0, m_BlendFinishedCount);
+            Assert.That(m_BlendManager.IsBlending, Is.True);
+
+            // change camera in the custom blend - we expect activation and deactivation events
+            customBlend.Blend.CamB = m_Cam3;
+            ProcessFrame(customBlend, 0.1f);
+            Assert.AreEqual(3, m_ActivatedEventCount);
+            Assert.AreEqual(1, m_DeactivatedEventCount);
+            Assert.AreEqual(1, m_BlendCreatedCount);
+            Assert.AreEqual(0, m_BlendFinishedCount);
+            Assert.That(m_BlendManager.IsBlending, Is.True);
+
+            customBlend.Blend.CamA = null;
+            ProcessFrame(customBlend, 1);
+            Assert.AreEqual(3, m_ActivatedEventCount);
+            Assert.AreEqual(2, m_DeactivatedEventCount);
             Assert.AreEqual(1, m_BlendCreatedCount);
             Assert.AreEqual(1, m_BlendFinishedCount);
             Assert.That(m_BlendManager.IsBlending, Is.False);